--- src/components/MusicPlayer/copyInputDiv.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import ReactPlayer from 'react-player';
-import PropTypes from 'prop-types';
-import musicPlayerUtils from './musicPlayerUtils';
-import mapStoreToProps from '../../redux/mapStoreToProps';
-
-class Application extends React.Component {
-  render() {
-    return (
-
-    );
-  }
-}
-
-export default Application;
